# docs/api/swagger.yaml
swagger: '2.0'
info:
  description: Entity service
  version: 1.0.0
  title: Well Delivery Service APIs
host: dns-host
basePath: /api/well-delivery
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  '/storage/v1/{type}':
    put:
      tags:
        - storage
      summary: Create or update entity
      description: >-
        The API represents the main injection mechanism into the Object
        Database. It allows entity creation and/or update. When no entity id is
        provided or when the provided id is not already present in the Object
        Database then a new entity is created. If the id is related to an
        existing entity in the Object Database then an update operation takes
        place and a new version of the entity is created. Required roles:
        service.storage.creator or service.storage.admin.
      operationId: Create or update entity
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: type
          in: path
          description: Entity type
          required: true
          type: string
          default: Well
        - name: Entity
          in: body
          required: true
          schema:
            $ref: '#/definitions/Entity'
      responses:
        '201':
          description: Entity created and/or updated successfully.
          schema:
            $ref: '#/definitions/EntityReturn'
        '400':
          description: Invalid entity format.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: Entity Id is invalid.
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
      security:
        - Bearer: []
  '/storage/v1/{type}/{id}':
    get:
      tags:
        - storage
      summary: Get latest version of entity
      description: >-
        This API returns the latest version of the given entity type and id.
        Required roles: service.storage.viewer, service.storage.creator or
        service.storage.admin.
      operationId: Get latest versio of entity
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: type
          in: path
          description: Entity type
          required: true
          type: string
          default: Well
        - name: id
          in: path
          description: Entity id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Entity retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturn'
        '400':
          description: Invalid entity type
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: 'Invalid entity type: invalid-well'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Entity not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find entity with id:
                b0aa7248-9392-49f4-9e7c-d6215a91547d
      security:
        - Bearer: []
    delete:
      tags:
        - storage
      summary: Delete entity
      description: >-
        The API performs a logical deletion of the given entity and all of its
        versions.   This operation can be reverted later. Required roles:
        service.storage.creator or service.storage.admin.
      operationId: Delete entity
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: type
          in: path
          description: Entity type
          required: true
          type: string
          default: Well
        - name: id
          in: path
          description: Entity id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '204':
          description: Entity deleted successfully.
        '400':
          description: Invalid entity type
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: 'Invalid entity type: invalid-well'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Entity not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find entity with id:
                b0aa7248-9392-49f4-9e7c-d6215a91547d
      security:
        - Bearer: []
  '/storage/v1/{type}/{id}/{version}':
    get:
      tags:
        - storage
      summary: Get a specific version of the given entity
      description: >-
        The API retrieves the specific version of the given entity type and id.
        Required roles: service.storage.viewer, service.storage.creator or
        service.storage.admin.
      operationId: Get a specific version of the given entity
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: type
          in: path
          description: Entity type
          required: true
          type: string
          default: Well
        - name: id
          in: path
          description: Entity id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
        - name: version
          in: path
          description: Entity version
          required: true
          type: integer
          default: 1596719439820
          format: int64
      responses:
        '200':
          description: Entity retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturn'
        '400':
          description: Invalid entity type
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: 'Invalid entity type: invalid-well'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Entity id or version not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find entity version with id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7_1601570839996
      security:
        - Bearer: []
    delete:
      tags:
        - storage
      summary: Delete a specific version of entity
      description: >-
        The API performs a logical deletion of the given entity type, id, and
        version number.This operation can be reverted later. Required roles:
        service.storage.creator or service.storage.admin.
      operationId: Delete a specific version of entity
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: type
          in: path
          description: Entity type
          required: true
          type: string
          default: Well
        - name: id
          in: path
          description: Entity id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
        - name: version
          in: path
          description: Entity version
          required: true
          type: integer
          default: 1596719439820
          format: int64
      responses:
        '204':
          description: Entity deleted successfully.
        '400':
          description: Invalid entity type
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: 'Invalid entity type: invalid-well'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Entity id or version not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find entity version with id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7_1601570839996
      security:
        - Bearer: []
  '/storage/v1/{type}/versions/{id}':
    get:
      tags:
        - storage
      summary: Get all entity version numbers
      description: >-
        The API returns a list containing all version numbers for the given
        entity type and id. Required roles: service.storage.viewer,
        service.storage.creator or service.storage.admin.
      operationId: Get all entity version  numbers
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: type
          in: path
          description: Entity type
          required: true
          type: string
          default: Well
        - name: id
          in: path
          description: Entity id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Entity version numbers retrieved successfully.
          schema:
            $ref: '#/definitions/VersionNumbers'
        '400':
          description: Invalid entity type
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: 'Invalid entity type: invalid-well'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Entity not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find entity with id:
                b0aa7248-9392-49f4-9e7c-d6215a91547d
      security:
        - Bearer: []
  '/storage/v1/{type}/{id}:purge':
    delete:
      tags:
        - storage
      summary: Purge entity
      description: >-
        The API performs a physical deletion of the given entity and all of its
        versions.  This operation cannot be undone. Required roles:
        service.storage.creator or service.storage.admin.
      operationId: Purge entity
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: type
          in: path
          description: Entity type
          required: true
          type: string
          default: Well
        - name: id
          in: path
          description: Entity id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '204':
          description: Entity purged successfully.
        '400':
          description: Invalid entity type
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: 'Invalid entity type: invalid-well'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Entity not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find entity with id:
                b0aa7248-9392-49f4-9e7c-d6215a91547d
      security:
        - Bearer: []
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  AppError:
    type: object
    properties:
      message:
        type: string
  Entity:
    type: object
    description: Entity object
    required:
      - kind
      - acl
      - legal
      - data
    properties:
      id:
        $ref: '#/definitions/EntityID'
      kind:
        $ref: '#/definitions/EntityKind'
      acl:
        $ref: '#/definitions/ACL'
      legal:
        $ref: '#/definitions/Legal'
      data:
        $ref: '#/definitions/EntityData'
  EntityReturn:
    type: object
    description: Entity return object
    properties:
      id:
        $ref: '#/definitions/EntityID'
      kind:
        $ref: '#/definitions/EntityKind'
      version:
        type: integer
        example: 1831253916104085
        format: int64
        description: The version number of this OSDU resource; set by the framework
      acl:
        $ref: '#/definitions/ACL'
      legal:
        $ref: '#/definitions/Legal'
      valid:
        type: boolean
        example: true
        description: The entity is valid or not.
      data:
        $ref: '#/definitions/EntityData'
...
  EntityID:
    type: string
    example: 'namespace:master-data--ExampleType:47c5967c-2961-55ea-96d7-e976a7954c26'
    description: The SRN which identifies this OSDU resource object without version.
    pattern: '^[\w\-\.]+:master-data\-\-ExampleType:[\w\-\.\:\%]+$'
  EntityKind:
    type: string
    example: 'slb-osdu-dev-des-prod-testing:well-delivery:exampletype:1.0.0'
    description: >-
      Kind of data it is being ingested and following the  pattern
      <Namespace>:<Source>:<Type>:<VersionMajor>.<VersionMinor>.<VersionPatch>. 
      The versioning scheme follows the semantic versioning,
      https://semver.org/..
    pattern: '^[\w\-\.]+:[\w\-\.]+:[\w\-\.]+:[0-9]+.[0-9]+.[0-9]+$'
  ACL:
    type: object
    description: The access control tags associated with this entity.
    required:
      - owners
      - viewers
    properties:
      viewers:
        type: array
        description: >-
          List of valid groups which will have view/read privileges over the
          entity.
        items:
          type: string
          example: data.default.viewers@slb-osdu-dev-des-prod-testing.contoso.com
      owners:
        type: array
        description: List of valid groups which will have write privileges over the entity.
        items:
          type: string
          example: data.default.owners@slb-osdu-dev-des-prod-testing.contoso.com
  Legal:
    type: object
    description: >-
      Attributes which represent the legal constraints associated with the
      record.
    properties:
      legaltags:
        type: array
        description: List of legaltag names associated with the entity.
        uniqueItems: true
        items:
          type: string
          example: sample-legal-tag
      otherRelevantDataCountries:
        type: array
        description: List of other relevant data countries.
        uniqueItems: true
        items:
          type: string
          example: US
          pattern: '^[A-Z]{2,2}$'
  VersionNumbers:
    type: object
    properties:
      id:
        type: string
        example: 374ba092-5a20-56f9-ada4-bc2cf71f2e50
      versions:
        type: array
        example:
          - 1597699007194
          - 1598021313729
          - 1598021463526
        items:
          type: integer
    description: Entity version numbers.
  EntityData:
    type: object
    description: Record payload represented as a list of key-value pairs.
    example:
      ExistenceKind: 'namespace:reference-data--ExistenceKind:planned:'
      NameWell: Well 01
      NameWellbore: Wellbore 01
...