# wd-core/src/main/java/org/opengroup/osdu/wd/core/models/EntityDto.java
// Copyright 2020 Schlumberger
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package org.opengroup.osdu.wd.core.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;


@Data
@AllArgsConstructor
@NoArgsConstructor
public class EntityDto {
    private String id;
    private String entityId;
    private String entityType;
    private String kind;
    private long version;
    private Legal legal;
    private ACL acl;
    private String existenceKind;
    private boolean valid;
    private String timeStamp;
    private String startTime;
    private String endTime;
    private Object data;
    private Object meta;
}